sync document updated: store raw data in document token; format on compile
